fix(link): dedup evidence must match what the linker can actually resolve
Three stacked defects in the UI-lib strip/dedup pipeline broke every macOS
UI doc-test link with 'Undefined symbols: <Stdout as Write>::flush' (#8455):

1. The trim counted libperry_stdlib.a symbols as 'provided elsewhere' even
   when stdlib is not on the final link line (pure-UI macOS programs link
   runtime + UI lib only). New StdlibEvidence::{Linked,NotLinked,Locate}
   threads the actual link composition into the trim and the localize pass;
   tier-3/COFF callers keep legacy self-location.

2. Evidence came from the members' nlist symbol tables, but the linker
   selects archive members exclusively via the archive INDEX (armap) — the
   two diverge for thin-LTO-internalized std CGUs and objcopy-rewritten
   members. collect_archive_global_symbols_flat now reads the armap
   (nm fallback only when an archive has no index), and the member-side
   subset test uses extern-defined symbols only (a local can never be
   referenced cross-object).

3. dedup_ui_lib_against_linked_libs localized bundled globals via
   llvm-objcopy --localize-symbols, which leaves LC_DYSYMTAB claiming
   nextdefsym=0 while N_EXT symbols remain in the nlist; llvm-ar then
   drops the member's surviving exports from the archive index and ld64
   cannot load it. Localization is now expressed as the inverse
   --keep-global-symbols keep-list, which rewrites the symbol table
   consistently.

Validated on a controlled repro (runtime archive whose index lacks the
flush symbol, matching CI's thin-LTO build): before, exactly CI's link
error; after, the snippet links and runs in both the degraded and the
pristine runtime configurations. strip_dedup unit tests moved to
strip_dedup/strip_dedup_tests.rs to stay under the 2000-line gate; new
regression test stdlib_evidence_gate_keeps_members_when_stdlib_is_not_linked
covers the NotLinked/Linked pair.

Fixes #8455
