+++ crates/perry/src/commands/compile.rs
@@ -112,7 +112,8 @@ use strip_dedup::{
     dedup_native_lib_for_tier3, dedup_runtime_for_tier3, dedup_stdlib_for_tier3,
     dedup_ui_lib_against_linked_libs, localize_stdlib_stub_symbols,
     strip_bundled_runtime_from_well_known_lib, strip_bundled_shared_deps_from_well_known_lib,
-    strip_duplicate_objects_from_lib, strip_duplicate_objects_from_well_known_lib,
+    strip_duplicate_objects_from_lib_with_evidence, strip_duplicate_objects_from_well_known_lib,
+    StdlibEvidence,
 };
 use targets::{
     apple_sdk_version, find_visionos_swift_runtime, find_watchos_swift_runtime,
