+++ crates/perry/src/commands/compile/link/mod.rs
@@ -38,9 +38,10 @@ use super::{
     find_perry_windows_sdk, find_stdlib_library, find_ui_library, find_visionos_swift_runtime,
     find_watchos_swift_runtime, is_android_target, is_native_windows_target, is_windows_target,
     localize_stdlib_stub_symbols, rust_target_triple, strip_bundled_runtime_from_well_known_lib,
-    strip_bundled_shared_deps_from_well_known_lib, strip_duplicate_objects_from_lib,
+    strip_bundled_shared_deps_from_well_known_lib, strip_duplicate_objects_from_lib_with_evidence,
     strip_duplicate_objects_from_well_known_lib, windows_pe_subsystem_flag,
-    windows_subsystem_needs_ui, windows_target_arch, CompilationContext, WindowsTargetArch,
+    windows_subsystem_needs_ui, windows_target_arch, CompilationContext, StdlibEvidence,
+    WindowsTargetArch,
 };
 
 mod archive_cache;
